+++ scripts/constants.js
@@ -1,0 +1,1 @@
+export const MODULE_ID = "animated-spell-effects-cartoon";